backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Milvus Vector Database
pymilvus==2.4.0
marshmallow<4.0.0  # Pin for compatibility with environs/pymilvus

# Google Gemini (LLM + Embeddings)
google-generativeai>=0.7.0

# Document processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Text processing
tiktoken==0.5.2

# Validation and utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.2.3
structlog==24.1.0
